Add presistent project directories
We are storing the retention string in the app's local store to
be used upon reopening the application. Currently we do not have
a way to remove this string within the application, resulting in
the application always opening the origional directory with no
way to change it. We also now have a identity and storage service.

Updates version to 0.0.2 pre-alpha.