Add inital terminal support & update context menus
Adds the terminal component to the project powered by xterms,
though it is not actually doing anything yet. Not sure if it should
use a native client plugin to access the terminal or a node proxy.
As for the context menu, this adds options to close, open projects,
and to close, save, and select tabs.