Fix issue with editor not able to open new files
When the user opened a file in the editor then created a new
document and opened it the editor bugged out and would show the
model of the last selected file instead of the new document. This
is fixed by checking if the model exists on passing in a file, and
if it is not we create it.