Fix issue with loading of modals at editor init
This adds a readOnly flag to the editor while we are loading the
models from the directory tree. This prevents the user from saving
a blank document while waiting for the editor to load. Additionally
this adds support for loading documents not found in the inital
load of models by creating them on the fly.